net/os-carp-vip-dhcp: split DhcpClient, policy and Keeper into leasekeeper/
WIP module split: extract Lease + DhcpClient (dhcpclient), ArpNudge +
FollowPolicy (policy) and Keeper + _identity_options (keeper) into the package.
All daemon code now lives under leasekeeper/; lease_keeper.py holds only the
entry points (argparse, logging, signals, main) plus transitional re-exports
for the tests. 123 tests green. Final commit removes the re-exports (tests
import from the package via a conftest facade), trims imports and lints 10/10.
