+++ net/os-carp-vip-dhcp/src/opnsense/scripts/OPNsense/CarpVipDhcp/lease_keeper.py
@@ -119,1100 +119,10 @@ from leasekeeper.codec import (  # noqa: F401  pylint: disable=unused-import
 
 
 
-@dataclass
-class Lease:
-    """The held DHCP binding: address, granting server, timing, and the
-    routing facts (options 3/1) the Parameter Request List asks for. During
-    a DORA it briefly holds the unconfirmed OFFER candidate; otherwise it is
-    the held (or last-held) binding. yiaddr None = unbound; expire() clears
-    only yiaddr, so the other fields survive as hints (logging, the
-    ARP-nudge target) until the next bind."""
-    yiaddr: str | None = None
-    server: str | None = None
-    lease_secs: int = DEFAULT_LEASE
-    t1_server: int | None = None   # server-provided renewal time (DHCP opt 58)
-    t2_server: int | None = None   # server-provided rebinding time (DHCP opt 59)
-    router: str | None = None      # default gateway (DHCP opt 3, fallback: server)
-    mask_bits: int | None = None   # subnet prefix length from opt 1
-
-
-class DhcpClient:  # pylint: disable=too-many-instance-attributes
-    """RFC 2131 client for one chaddr: owns the held binding (a Lease)
-    and the stateful protocol sequences
-    (INIT-REBOOT / DORA / RENEW / REBIND / RELEASE) with their send/await
-    machinery. It does NOT own the packet capture: sends travel through the
-    injected capture backend, and the capture owner hands xid-matched replies
-    to feed(). Policy stays with the caller through three
-    injected hooks: should_stop (abort mid-sequence), ensure_sniffer (capture
-    must be alive before a send), and on_changed_address(got, rx, phase,
-    release_on_enforce) -> bool for an ACK whose address differs from the
-    expected one (True = the caller validated and adopted it, see adopt())."""
-
-    def __init__(self, capture, chaddr, eth_src, id_opts, *,  # pylint: disable=too-many-arguments
-                 should_stop, ensure_sniffer, on_changed_address):
-        self._capture = capture
-        self.chaddr = chaddr
-        self.chraw = mac2raw(chaddr)
-        self.eth_src = eth_src
-
-        # Optional DHCP request options (empty -> not sent); added to every
-        # DISCOVER/REQUEST/RENEW so the server sees a consistent client identity.
-        self._id_opts = id_opts
-        self._should_stop = should_stop
-        self._ensure_sniffer = ensure_sniffer
-        self._on_changed = on_changed_address
-
-        self.xid = _new_xid()
-        self.binding = Lease()         # the held (or last-held) DHCP binding
-        self._tried_reboot = False     # the first acquire tries INIT-REBOOT before a full DISCOVER
-
-        self._rx = None                # latest DhcpReply snapshot (set via feed(), sniffer thread)
-        self._ev = threading.Event()
-
-    def feed(self, rx):
-        """Hand a first-party (xid-matched) DhcpReply to the waiting sequence.
-        Called on the sniffer thread; a lone ref-assign plus the event set."""
-        self._rx = rx
-        LOG.debug("DHCP reply: %s", _fmt_reply(rx))
-        self._ev.set()
-
-    def _send_dhcp(self, mtype, extra, ciaddr="0.0.0.0"):
-        # ciaddr is set for RENEW/REBIND (the client already owns the address);
-        # the broadcast flag stays on so the reply is reliably captured by the sniffer.
-        self._capture.send_dhcp(
-            eth_src=self.eth_src, ip_src=ciaddr, ip_dst=IPV4_BROADCAST,
-            chaddr=self.chraw, xid=self.xid, ciaddr=ciaddr, flags=BROADCAST_FLAG,
-            options=_dhcp_options(mtype, extra, self._id_opts))
-
-    def _wait_for_dhcp_reply(self, want, timeout):
-        """Wait up to timeout for a reply of message-type `want`. Returns the
-        DhcpReply on match, the string "NAK" on DHCPNAK, or None on timeout.
-        Returning the snapshot avoids re-reading self._rx (set by the sniffer
-        thread) after the wait."""
-        end = time.time() + timeout
-        while time.time() < end and not self._should_stop():
-            self._ev.clear()
-            self._ev.wait(min(1.0, max(0.05, end - time.time())))
-
-            rx = self._rx
-            if rx and rx.mtype == want:
-                return rx
-            if rx and rx.mtype == NAK:
-                # Surface the server's option-56 text (why it refused) if present;
-                # it is the operator's main clue for a rejected renew.
-                txt = _msg_text(rx.message)
-                reason = f" -- {txt}" if txt else ""
-                LOG.warning("DHCPNAK received (server %s, xid 0x%08x%s)%s",
-                            rx.server_id or "unknown", self.xid,
-                            f" via relay {rx.giaddr}" if rx.giaddr else "", reason)
-                return "NAK"
-        return None
-
-    def _absorb_reply(self, rx, default_lease):
-        """Adopt lease timing, gateway (opt 3) and subnet mask (opt 1) from an ACK
-        -- the one place that knows which DhcpReply fields carry lease state.
-        gateway + mask are what the Parameter Request List asks for; keep them for
-        logging + the cross-subnet follow decision. The lease time is floored at
-        MIN_LEASE so a server (or a spoofed ACK) offering a 1-second lease cannot
-        drive the renew loop into a tight broadcast/CPU spin."""
-        self.binding.lease_secs = max(rx.lease or default_lease, MIN_LEASE)
-        self.binding.t1_server, self.binding.t2_server = rx.t1, rx.t2
-        self.binding.router = rx.router or self.binding.router
-        self.binding.mask_bits = _mask_to_bits(rx.subnet_mask) or self.binding.mask_bits
-
-    def adopt(self, rx):
-        """Bind to the address in rx: the ACK that completes a DORA, or a
-        validated changed address the caller's follow policy accepted (the
-        on_changed_address hook calls this, so the lease state stays owned here
-        while the decision stays with the caller). The server refreshes from
-        the ACK's server-id, keeping the previous one when the ACK has none."""
-        self.binding.yiaddr = rx.yiaddr
-        self.binding.server = rx.server_id or self.binding.server
-        self._absorb_reply(rx, DEFAULT_LEASE)
-
-    def expire(self):
-        """Forget the held binding WITHOUT a RELEASE (on-time lease expiry, or a
-        grant the policy refused to keep): yiaddr clears so the next cycle
-        re-acquires via DISCOVER; server/router stay as hints for logging and
-        the caller's nudge target."""
-        self.binding.yiaddr = None
-
-    def acquire(self, request_ip):
-        """Get a lease. The first acquire after (re)start tries INIT-REBOOT (a
-        direct REQUEST for our known address) before a full DISCOVER; after that,
-        the normal DORA. INIT-REBOOT is one exchange and surfaces a NAK when the
-        server is reachable but refuses the address.
-
-        Startup-only (the `_tried_reboot` latch): a genuine on-time lease expiry
-        re-acquires via DISCOVER (RFC 2131 4.4.5), NOT INIT-REBOOT. We do not track
-        lease expiry across a restart -- the server arbitrates the requested address
-        (ACK if still ours, NAK/silence -> fall back to DISCOVER)."""
-        if not self._tried_reboot:
-            self._tried_reboot = True
-            if request_ip:
-                if self.reboot(request_ip):
-                    return True
-                LOG.info("INIT-REBOOT did not bind %s -- falling back to DISCOVER", request_ip)
-
-        return self.dora(request_ip)
-
-    def dora(self, request_ip=None):
-        """Acquire a lease via the DHCP DORA handshake: the SELECTING phase
-        (DISCOVER until an OFFER arrives) then the REQUESTING phase (REQUEST
-        the offer until the ACK lands). Returns True once BOUND, False on
-        failure/NAK."""
-        self.xid = _new_xid()
-        return self._discover(request_ip) and self._request_offer(request_ip)
-
-    def _discover(self, request_ip):
-        """SELECTING: broadcast DISCOVER until a server OFFERs; record the
-        offered address and server. False on NAK/timeout/stop."""
-        extra = [("requested_addr", request_ip)] if request_ip else []
-        for attempt in range(1, DORA_ATTEMPTS + 1):
-            if self._should_stop():
-                return False
-            self._ensure_sniffer()
-            self._rx = None
-            try:
-                self._send_dhcp("discover", extra)
-            except Exception as e:
-                LOG.error("DHCP DISCOVER send failed: %s", e)
-                time.sleep(SEND_RETRY_DELAY)
-                continue
-            rx = self._wait_for_dhcp_reply(OFFER, REPLY_TIMEOUT)
-            if rx == "NAK":
-                return False
-            if rx:
-                self.binding.yiaddr, self.binding.server = rx.yiaddr, rx.server_id
-                return True
-            # xid included so the exchange can be matched against a packet capture.
-            LOG.info("no DHCP OFFER (attempt %d, xid 0x%08x)", attempt, self.xid)
-            time.sleep(min(_jittered(2 ** attempt), ATTEMPT_BACKOFF_CAP))
-        return False
-
-    def _request_offer(self, request_ip):
-        """REQUESTING: REQUEST the offered address until the ACK lands and we
-        are BOUND. False on NAK/timeout/stop; an ACK for a different address
-        goes through the on_changed_address policy hook."""
-        for attempt in range(1, DORA_ATTEMPTS + 1):
-            if self._should_stop():
-                return False
-            self._ensure_sniffer()   # an interface flap between OFFER and here would leave us deaf
-            self._rx = None
-            try:
-                self._send_dhcp(
-                    "request",
-                    [("server_id", self.binding.server), ("requested_addr", self.binding.yiaddr)],
-                )
-            except Exception as e:
-                LOG.error("DHCP REQUEST send failed: %s", e)
-                time.sleep(SEND_RETRY_DELAY)
-                continue
-            rx = self._wait_for_dhcp_reply(ACK, REPLY_TIMEOUT)
-            if rx == "NAK":
-                return False   # NAK -> back to INIT; the run loop re-acquires (DISCOVER)
-            if rx:
-                got = rx.yiaddr
-                if request_ip and got != request_ip:
-                    return self._on_changed(got, rx, PHASE_DORA, True)
-                self.adopt(rx)
-                return True
-            LOG.info("no DHCP ACK from %s for %s (attempt %d, xid 0x%08x)",
-                     self.binding.server, self.binding.yiaddr, attempt, self.xid)
-            time.sleep(min(_jittered(2 ** attempt), ATTEMPT_BACKOFF_CAP))
-        return False
-
-    # reboot() and renew() share _request_offer's REQUEST->ACK shape: they are
-    # the RFC 2131 REBOOTING and RENEWING/REBINDING states next to REQUESTING.
-    # They stay separate linear loops on purpose -- the three differ on nine
-    # axes (options, ciaddr, attempts, timeout, backoff, send-failure policy,
-    # expected address, changed-address phase, bind action), so a shared
-    # parametrized engine would turn each into config that is harder to audit
-    # against the RFC than the plain loop it replaces.
-    def reboot(self, request_ip):
-        """INIT-REBOOT (RFC 2131 4.3.2): we already know the address we want
-        (request_ip), so REQUEST it directly instead of a full DISCOVER -- one
-        exchange, and a server that is reachable but refuses the address answers
-        with a NAK (a DISCOVER it may just ignore), surfacing 'reachable but
-        refused' in the log. server_id MUST NOT be set and ciaddr stays 0 (we do
-        not own the address on the interface). Returns True once BOUND, False on
-        NAK/timeout so the caller falls back to a full DORA."""
-        if not request_ip:
-            return False
-
-        self.xid = _new_xid()
-        extra = [("requested_addr", request_ip)]
-        for attempt in range(1, REBOOT_ATTEMPTS + 1):
-            if self._should_stop():
-                return False
-            self._ensure_sniffer()
-            self._rx = None
-            try:
-                self._send_dhcp("request", extra)
-            except Exception as e:
-                LOG.error("DHCP INIT-REBOOT send failed: %s", e)
-                time.sleep(SEND_RETRY_DELAY)
-                continue
-            rx = self._wait_for_dhcp_reply(ACK, REPLY_TIMEOUT)
-            if rx == "NAK":
-                return False   # server refused our known address -> full DISCOVER
-            if rx:
-                got = rx.yiaddr
-                if got and got != request_ip:
-                    return self._on_changed(got, rx, PHASE_REBOOT, True)
-                self.adopt(rx)
-                if not self.binding.yiaddr:
-                    self.binding.yiaddr = request_ip   # ACK without yiaddr: we asked for it
-                return True
-            LOG.info("no DHCP ACK to INIT-REBOOT for %s (attempt %d, xid 0x%08x)",
-                     request_ip, attempt, self.xid)
-            if attempt < REBOOT_ATTEMPTS:   # no backoff after the last try -- fall to DORA at once
-                time.sleep(min(_jittered(2 ** attempt), ATTEMPT_BACKOFF_CAP))
-        return False
-
-    # One return per protocol gate (unbound/stop/send-fail/NAK/changed/ok);
-    # folding them into an attempt-helper trades readable gates for plumbing.
-    def renew(self, rebind=False):  # pylint: disable=too-many-return-statements
-        """REQUEST an extension of the held lease. Returns True on a fresh ACK,
-        False on NAK/timeout/unbound (the caller escalates: RENEW -> REBIND ->
-        re-acquire)."""
-        # RENEWING/REBINDING (RFC 2131 4.3.2): ciaddr identifies the lease, so
-        # server_id AND requested_addr MUST NOT be set in either state. We always
-        # broadcast (the co-resident non-promiscuous sniffer needs the reply
-        # broadcast), so any server may answer, which is fine for a single-server
-        # ISP WAN. `rebind` is still load-bearing: it picks the log label AND
-        # (via the `phase` it sets) relaxes the expected-server check in the
-        # caller's on_changed_address hook, since at T2 any server may
-        # legitimately answer.
-        yiaddr = self.binding.yiaddr
-        if not yiaddr:
-            return False   # nothing bound -> nothing to renew
-
-        opts = []
-        for _ in range(RENEW_ATTEMPTS):
-            if self._should_stop():
-                return False
-            self._ensure_sniffer()
-            self._rx = None
-            try:
-                self._send_dhcp("request", opts, ciaddr=yiaddr)
-            except Exception as e:
-                LOG.error("DHCP %s send failed: %s", PHASE_REBIND if rebind else PHASE_RENEW, e)
-                return False
-            rx = self._wait_for_dhcp_reply(ACK, RENEW_TIMEOUT)
-            if rx == "NAK":
-                return False   # NAK -> re-DORA
-            if rx:
-                got = rx.yiaddr
-                if got and got != self.binding.yiaddr:
-                    # Some dynamic servers change the address at renewal (ACK with a
-                    # new yiaddr) instead of NAKing. Route it through the same
-                    # follow / enforce decision (and hardening) as the initial DORA.
-                    phase = PHASE_REBIND if rebind else PHASE_RENEW
-                    return self._on_changed(got, rx, phase, False)
-                self._absorb_reply(rx, self.binding.lease_secs)
-                return True
-        return False
-
-    def release(self, yiaddr=None, server=None):
-        """RELEASE a lease -- the held one by default, or an explicit
-        (yiaddr, server) pair (the enforce path releases an address it was
-        granted but refuses to keep)."""
-        if yiaddr is None:
-            yiaddr, server = self.binding.yiaddr, self.binding.server
-        if not yiaddr:
-            return
-
-        try:
-            # No broadcast flag: RELEASE expects no reply to capture.
-            self._capture.send_dhcp(
-                eth_src=self.eth_src, ip_src=yiaddr, ip_dst=server or IPV4_BROADCAST,
-                chaddr=self.chraw, xid=self.xid, ciaddr=yiaddr, flags=0,
-                options=[("message-type", "release"), ("server_id", server), "end"])
-            LOG.info("DHCP RELEASE of lease %s sent (server %s)", yiaddr, server or "broadcast")
-        except Exception as e:
-            LOG.error("RELEASE failed: %s", e)
-
-    def timing(self):
-        """Effective renew (T1) / rebind (T2) seconds and where they came from.
-
-        Uses server-provided DHCP option 58/59 when present and sane, otherwise
-        the RFC-suggested 0.5 / 0.875 of the lease time.
-        """
-        lease = max(1, self.binding.lease_secs)
-        t1 = self.binding.t1_server if self.binding.t1_server else int(lease * T1_FACTOR)
-        t2 = self.binding.t2_server if self.binding.t2_server else int(lease * T2_FACTOR)
-        # Keep both timers inside the lease; only apply the MIN_T1 floor when the
-        # lease is long enough to accommodate it (very short leases renew sooner).
-        t1 = min(t1, lease)
-        if lease > MIN_T1:
-            t1 = max(MIN_T1, t1)
-        t2 = min(max(t1 + REBIND_MARGIN, t2), lease)
-        src = "server" if (self.binding.t1_server or self.binding.t2_server) else "derived"
-        return t1, t2, src
-
-
-class ArpNudge:  # pylint: disable=too-many-instance-attributes
-    """Keep the upstream gateway's ARP entry for the leased address fresh, for
-    gateways that ignore gratuitous ARP and never re-ARP an expired entry (see
-    the README's "ARP nudge" section). Owns the nudge pacing and the gateway
-    reachability stamp; the caller supplies the current lease binding
-    (yiaddr, gateway) per call and a CARP-role probe at construction -- never
-    nudge from a backup (it would steal the VIP's traffic), so anything but a
-    confirmed MASTER fails closed (the next interval retries)."""
-
-    def __init__(self, capture, chaddr, interval, is_master):
-        self._capture = capture
-        self.chaddr = chaddr
-        # Interval floor so a typo cannot flood the segment; 0 = disabled.
-        self.interval = max(ARP_NUDGE_MIN, interval) if interval else 0
-        self._is_master = is_master    # callable -> True/False/None (None = probe failed)
-
-        self.last_nudge = 0.0          # epoch of the last sent nudge (0 = never)
-        # Reachability: the sniffer stamps last_reply when the gateway answers
-        # a nudge (a lone atomic float write); the status page surfaces its age.
-        self.last_reply = 0.0          # epoch of the gateway's last ARP reply (0 = none)
-        self._gw = None                # last nudge target we logged (log again on change)
-        self._warned = False           # warned once about a missing nudge target
-
-    def maybe_nudge(self, yiaddr, gateway, force=False):
-        """Refresh the gateway's ARP entry for yiaddr by broadcasting an ARP
-        request from (yiaddr, chaddr). No-op unless enabled, bound, due (or
-        forced) and CARP master."""
-        if not self.interval or not yiaddr:
-            return
-        if not force and time.time() - self.last_nudge < self.interval:
-            return
-        if self._is_master() is not True:
-            return
-
-        if not gateway:
-            # Enabled but no target: without this warning the nudge would be a
-            # silent no-op and the operator would believe they are protected.
-            if not self._warned:
-                LOG.warning("ARP nudge enabled but no gateway known "
-                            "(no DHCP router option or server-id) -- cannot nudge")
-                self._warned = True
-            return
-
-        try:
-            # This frame is shaped to satisfy three ISP access-network guards at
-            # once (README "Playing nicely" section):
-            #   * op=1 (a REQUEST, not a gratuitous announcement) -- gear that
-            #     filters unsolicited/gratuitous ARP still processes a request,
-            #     which is what refreshes the entry;
-            #   * sender (psrc=leased IP, hwsrc=chaddr=CARP MAC) is exactly the
-            #     DHCP-snooped IP<->MAC binding, so Dynamic ARP Inspection passes it;
-            #   * sending it at all exists for gateways that never re-ARP an
-            #     expired entry ("secured ARP") and would otherwise blackhole the
-            #     VIP. Only the CARP master sends it (gated above).
-            self._capture.send_arp_request(self.chaddr, yiaddr, gateway)
-            self.last_nudge = time.time()
-            # Log the first nudge (and any target change) at INFO so the default log
-            # shows the nudge is active; routine repeats at DEBUG (they fire oftener
-            # than DHCP renews). Whether the gateway actually answered is surfaced by
-            # age on the status page (via last_reply -> the heartbeat's arpok=).
-            if gateway != self._gw:
-                LOG.info("ARP nudge active: who-has %s tell %s (src %s) every %ds",
-                         gateway, yiaddr, self.chaddr, self.interval)
-                self._gw = gateway
-            else:
-                LOG.debug("ARP nudge sent: who-has %s tell %s", gateway, yiaddr)
-        except Exception as e:
-            LOG.warning("ARP nudge failed (target %s): %s", gateway, e)
-
-    def on_arp_reply(self, frame, yiaddr, gateway):
-        """Stamp last_reply when the gateway answers our nudge -- a reachability
-        signal the status page surfaces by age. Only the reply to OUR who-has
-        counts: op=2 (is-at), sender = the nudge target gateway, target = our
-        leased IP. Runs on the capture thread; the stamp is a lone atomic write.
-        Advisory only (an on-segment attacker could forge or withhold it);
-        nothing here feeds lease/CARP/follow decisions."""
-        if frame.op != 2:                       # 2 = is-at (reply); requests are filtered out in BPF
-            return
-        if not gateway or not yiaddr:
-            return
-        if frame.psrc == gateway and frame.pdst == yiaddr:
-            self.last_reply = time.time()
-            LOG.debug("ARP reply from %s (is-at) for %s", gateway, yiaddr)
-
-
-class FollowPolicy:  # pylint: disable=too-many-instance-attributes
-    """Decides what happens when the server grants a DIFFERENT address than
-    the target this keeper exists to hold. In follow mode: validate the new
-    address (plausibility, routability class, expected server), throttle
-    against flap/spoof storms, drive the configd VIP rewrite and adopt it
-    into the DHCP client. Otherwise (enforce, a fixed reservation): alarm,
-    release the refused grant and stay unbound. Owns the target address
-    (rewritten on a successful follow), the persisted follow throttle, the
-    apply-retry watchdog and the peer-ACK observation handoff."""
-
-    def __init__(self, target, follow, chaddr, dhcp, hb_mismatch):
-        self.target = target           # the address this keeper is meant to hold
-        self.follow = follow
-        self._dhcp = dhcp
-        self._hb_mismatch = hb_mismatch
-
-        self._followed_ip = None       # last address we asked configd to follow to
-        self._follow_from = None       # address we followed FROM (for the retry watchdog)
-        self._fired_at = 0.0           # when we last dispatched follow_update (retry deadline)
-        self._gw_args = []             # extra follow_update args on a cross-subnet move: [old_gw, new_gw, bits]
-
-        # Throttle state, keyed by chaddr so it survives the follow-induced
-        # restart (the target-keyed runtime paths change on every follow).
-        self._state_file = f"/var/run/carpvipdhcp-follow-{_fs_safe(chaddr)}"
-
-        # A changed ISP address first seen in the PEER's ACK (both HA nodes run
-        # an identical keeper on the same shared chaddr). The sniffer thread
-        # records it via observe() -- a lone atomic ref-assign -- and the main
-        # thread consumes it in check_observed().
-        self._observed = None
-
-    def _last_follow_time(self):
-        """Epoch of this chaddr's last follow (persisted so the throttle survives
-        the follow-induced restart). 0 if never / unreadable."""
-        try:
-            with open(self._state_file, encoding="utf-8") as f:
-                return float(f.read().strip())
-        except (OSError, ValueError):
-            return 0.0
-
-    def _record_follow(self):
-        try:
-            _atomic_write(self._state_file, str(int(time.time())))
-        except OSError as e:
-            LOG.warning("could not persist follow timestamp: %s", e)
-
-    def on_changed_address(self, got, rx, phase, release_on_enforce):
-        """An ACK arrived whose address differs from the one we hold/request.
-
-        In follow mode: validate the address (sane, same routability class, from
-        the expected server), throttle against flap/spoof storms, then adopt it
-        (rewrite the CARP VIP). Otherwise (enforce): alarm on the mismatch.
-        Returns True if we are now bound to `got`, False if the caller should
-        re-acquire.
-        """
-        if self.follow:
-            if not _sane_ipv4(got):
-                LOG.error("%s: ACK yiaddr %r from server %s implausible -- not following",
-                          phase, got, rx.server_id)
-                return False
-            if not _same_ip_class(self.target, got):
-                LOG.error("%s: refusing to follow %s -> %s across address class "
-                          "(possible spoofed ACK from %s)", phase, self.target, got,
-                          rx.server_id)
-                return False
-            leased_from = self._dhcp.binding.server
-            if phase != PHASE_REBIND and rx.server_id and leased_from and rx.server_id != leased_from:
-                LOG.error("%s: ACK from unexpected server %s (leased from %s) -- not following",
-                          phase, rx.server_id, leased_from)
-                return False
-
-            waited = time.time() - self._last_follow_time()
-            if waited < MIN_FOLLOW_INTERVAL:
-                LOG.warning("%s: follow %s -> %s throttled (%.0fs < %ds) -- deferring",
-                            phase, self.target, got, waited, MIN_FOLLOW_INTERVAL)
-                return False
-
-            LOG.warning("ISP gave %s (VIP was %s) at %s -- following: updating the CARP VIP",
-                        got, self.target, phase)
-            # If the ISP also moved the gateway (a cross-subnet renumber), follow the
-            # new gateway + prefix too so outbound keeps working -- parity with what a
-            # plain DHCP interface does. Needs the new subnet mask (opt 1) to set the
-            # VIP prefix; without it we can only move the address, so warn instead.
-            self._gw_args = []
-            old_router = self._dhcp.binding.router
-            if rx.router and old_router and rx.router != old_router:
-                bits = _mask_to_bits(rx.subnet_mask)
-                if bits:
-                    LOG.warning("follow %s -> %s also moves the gateway (%s -> %s), subnet "
-                                "/%d -- following across the subnet", self.target, got,
-                                old_router, rx.router, bits)
-                    self._gw_args = [old_router, rx.router, str(bits)]
-                else:
-                    LOG.error("follow %s -> %s changes the gateway (%s -> %s) but the ACK "
-                              "carried no subnet mask -- updating the VIP address only; fix the "
-                              "interface prefix + System->Gateways by hand", self.target,
-                              got, old_router, rx.router)
-
-            self._record_follow()
-            # _follow_update reads target as the old address, so remember it
-            # (for the retry watchdog) and fire before overwriting target.
-            self._follow_from = self.target
-            self._follow_update(got)
-            self.target = got
-            self._dhcp.adopt(rx)
-            return True
-        # Enforce: a fixed reservation must always return the target.
-        LOG.error("%s: IP mismatch -- server %s gave %s, requested %s (reservation problem?)",
-                  phase, rx.server_id, got, self.target)
-        self._hb_mismatch(got, self.target)
-        if release_on_enforce:
-            self._dhcp.release(got, rx.server_id)
-            # DORA's OFFER phase already recorded a tentative yiaddr; drop it so
-            # the run loop re-acquires instead of renewing the refused grant.
-            self._dhcp.expire()
-        return False
-
-    def _follow_update(self, new_ip):
-        """Ask configd to rewrite the CARP VIP (and this keeper's reference) from
-        the target to new_ip, then reconfigure. Fire-and-forget: the resulting
-        service restart replaces this daemon with one bound to the new address."""
-        if new_ip == self._followed_ip:
-            return   # already asked for this address
-        try:
-            self._fire_follow_update(self.target, new_ip)
-            # Only mark as handled once the request was actually dispatched, so a
-            # spawn failure is retried next cycle instead of getting stuck.
-            self._followed_ip = new_ip
-            LOG.info("requested CARP VIP update %s -> %s", self.target, new_ip)
-        except Exception as e:
-            LOG.error("follow_update request failed: %s", e)
-
-    def _fire_follow_update(self, old_ip, new_ip):
-        """Dispatch the configd follow_update action (old -> new) and stamp the
-        retry deadline. Separate from _follow_update so the watchdog can re-drive
-        a stalled follow without tripping the _followed_ip equality guard."""
-        cmd = ["/usr/local/sbin/configctl", "-d", "carpvipdhcp", "follow_update", old_ip, new_ip]
-        cmd += self._gw_args   # cross-subnet: [old_gw, new_gw, bits]; empty on a same-subnet move
-        subprocess.Popen(cmd)  # pylint: disable=consider-using-with
-        self._fired_at = time.time()
-
-    def watchdog(self):
-        """Re-drive a follow that never took effect. After a successful follow,
-        follow_update restarts this daemon within a few seconds; if we are still
-        alive well past FOLLOW_RETRY_DEADLINE, its apply failed or stalled, so
-        re-dispatch it (idempotent: it reconverges whether the config already
-        moved or not, and its rc.d restart <old-id> eventually replaces us)."""
-        if not (self.follow and self._followed_ip and self._follow_from):
-            return
-        if time.time() - self._fired_at < FOLLOW_RETRY_DEADLINE:
-            return
-        LOG.warning("follow %s -> %s not applied within %ds -- re-driving",
-                    self._follow_from, self._followed_ip, FOLLOW_RETRY_DEADLINE)
-        try:
-            self._fire_follow_update(self._follow_from, self._followed_ip)
-        except Exception as e:
-            LOG.error("follow_update retry failed: %s", e)
-
-    def observe(self, rx):
-        """Record a peer-ACK observation (capture thread; a lone ref-assign);
-        the main loop consumes it in check_observed().
-
-        The read-then-clear in check_observed is not atomic against this store,
-        so a peer ACK landing in that narrow window can be dropped. Accepted:
-        the next peer ACK or our own renewal re-detects the change, and the 1s
-        maintain tick keeps polling -- convergence is delayed, never lost."""
-        self._observed = rx
-
-    def check_observed(self):
-        """Adopt an address change first seen in the PEER's DHCP ACK (same shared
-        chaddr) without waiting for our own renewal timer. The sniffer only
-        records the observation; the follow itself runs here on the main thread,
-        through the same hardening/throttle as a first-party ACK. This collapses
-        the follow window that would otherwise leave the two nodes on different
-        VIP prefixes long enough for the backup to promote (transient
-        dual-master; see docs/single-ip-wan-carp.md section 3)."""
-        rx = self._observed
-        if rx is None:
-            return
-        self._observed = None
-        if not self.follow or not rx.yiaddr or rx.yiaddr == self._dhcp.binding.yiaddr:
-            return
-        LOG.info("observed peer DHCP ACK for %s (we hold %s) -- following early to converge",
-                 rx.yiaddr, self._dhcp.binding.yiaddr)
-        self.on_changed_address(rx.yiaddr, rx, PHASE_OBSERVED, release_on_enforce=False)
-
-
-class Keeper:  # pylint: disable=too-many-instance-attributes
-    """Orchestration around the components: owns the capture backend (feeding
-    DHCP replies to DhcpClient, ARP replies to ArpNudge and peer-ACK
-    observations to FollowPolicy), the heartbeat, the CARP role watch, the
-    acquire pacing (backoff + link-return fast path) and the signal-driven
-    operator actions. The lease lives in DhcpClient; the changed-address
-    decision and the target address live in FollowPolicy."""
-
-    def __init__(self, iface, chaddr, request_ip=None, eth_src=None, *,  # pylint: disable=too-many-arguments
-                 hbfile=None, release_on_exit=False, vhid=None,
-                 follow=False, vendor_class=None, client_id=None, hostname=None,
-                 arp_nudge=0, arp_listen_promisc=False, capture_backend="scapy"):
-        self.iface = iface
-        self.chaddr = chaddr.lower()
-        self.eth_src = (eth_src or chaddr).lower()
-        self.hbfile = hbfile
-        self.release_on_exit = release_on_exit
-        self.vhid = str(vhid) if vhid else None
-
-        # Capture backend component: owns the capture socket/thread and the
-        # wire codec on both directions, and hands decoded neutral frames
-        # back (DHCP replies and ARP replies, routed below). Promiscuous
-        # capture is off by default; the rationale lives in the module
-        # docstring's security section.
-        self._capture = CAPTURE_BACKENDS[capture_backend](
-            iface, arp_listen_promisc, self._on_dhcp_reply, self._on_arp_reply)
-
-        # ARP nudge component: owns the pacing and reachability state; the
-        # keeper supplies the lease binding per nudge (_arp_nudge) and the
-        # CARP-role probe (via the _carp_master_probe shim).
-        self._nudge = ArpNudge(self._capture, self.chaddr, arp_nudge, self._carp_master_probe)
-
-        self._was_master = None        # CARP role at the last nudge check (None = unknown yet)
-        self._nudge_now = False        # operator asked for an immediate nudge (SIGUSR1)
-        self._poll_role_now = False    # a CARP transition fired -> re-check role now (SIGUSR2)
-        self._renew_asap = False       # renew at the next _hold_lease tick instead of waiting for T1
-
-        # DHCP client component: owns the lease state and the protocol sequences;
-        # the keeper owns the capture (feeding it xid-matched replies) and the
-        # policy hooks.
-        self._dhcp = DhcpClient(
-            self._capture, self.chaddr, self.eth_src,
-            _identity_options(vendor_class, client_id, hostname),
-            should_stop=lambda: self._stop,
-            ensure_sniffer=self._ensure_sniffer,
-            on_changed_address=self._on_changed_address)
-
-        # Follow/enforce policy component: owns the target address and the
-        # follow bookkeeping; drives the client via adopt()/release()/expire().
-        self._follow = FollowPolicy(request_ip, follow, self.chaddr, self._dhcp,
-                                    self._hb_mismatch)
-
-        self.redora_wait = REDORA_MIN
-        # Link-return fast path (only while UNBOUND): a carrier down->up edge resets
-        # the backoff and re-DORAs at once, like dhclient's link-up -> state_reboot.
-        self._link_up = None           # last carrier state seen (None = unknown / not probed)
-        self._link_kick_at = 0.0       # epoch of the last link-return kick (debounce)
-        self._link_returned = False    # set by _sleep_interruptible on a carrier return while unbound
-
-        self._stop = False
-        # General early-wake for the maintain-loop sleep (_sleep_interruptible): lets it
-        # return before the 1s tick when there is pending work. Currently the
-        # sniffer sets it on an observed address change so the follow fires in
-        # milliseconds; a future fast-wake need should reuse this rather than mint
-        # a second event. Set only by the capture thread; waited/cleared only by
-        # the main thread.
-        self._wake = threading.Event()
-
-    # ---- capture (resilient) ----
-    def _ensure_sniffer(self):
-        if not self._capture.alive():
-            LOG.warning("DHCP-reply capture down -- (re)starting")
-            self._capture.start()
-            time.sleep(1)
-
-    def _on_arp_reply(self, frame):
-        """Capture-backend ARP callback: the gateway answering our nudge
-        (reachability stamp); the nudge component does the matching."""
-        self._nudge.on_arp_reply(frame, self._dhcp.binding.yiaddr, self._nudge_target())
-
-    def _chaddr_matches(self, frame):
-        """True if the reply's BOOTP client hardware address is our chaddr (the
-        CARP virtual MAC). Used to accept the PEER's ACK on the shared chaddr:
-        the peer node runs an identical keeper on the very same chaddr."""
-        return frame.chaddr[:6] == self._dhcp.chraw
-
-    def _on_dhcp_reply(self, frame):
-        # Runs on the capture thread under the backend's handler guard
-        # (_deliver), so an unexpected failure here is logged and dropped there.
-        if frame.op != BOOTREPLY:
-            return
-        # First-party path: a reply to OUR in-flight exchange (random xid,
-        # regenerated per DORA). Parsed and fed to the waiting client sequence.
-        if frame.xid == self._dhcp.xid:
-            self._dhcp.feed(_parse_reply(frame))
-            return
-        # Not our xid. In follow mode, still watch for the PEER's ACK: both HA
-        # nodes run an identical keeper on the SAME chaddr (the CARP virtual
-        # MAC), so the peer's ACK reveals a changed ISP address one exchange
-        # sooner than our own renewal timer -- closing the window where the
-        # two nodes hold different VIP prefixes long enough for CARP to
-        # dual-master (see docs/single-ip-wan-carp.md, section 3). This is a
-        # deliberately narrow relaxation of the xid trust gate: it requires
-        # our chaddr and an ACK for a plausible, different address, and it
-        # only RECORDS the observation (one atomic ref write) for the main
-        # thread -- which routes it through the same follow hardening
-        # (sane / same-class / expected-server / throttle) as a first-party
-        # ACK and never acts on the capture thread.
-        if not self._follow.follow or not self._dhcp.binding.yiaddr or not self._chaddr_matches(frame):
-            return
-        rx = _parse_reply(frame)
-        if (rx.mtype == ACK and rx.yiaddr and rx.yiaddr != self._dhcp.binding.yiaddr
-                and _sane_ipv4(rx.yiaddr)):
-            self._follow.observe(rx)
-            self._wake.set()   # wake the maintain-loop sleep now, don't wait for the tick
-
-    # ---- heartbeat / status file ----
-
-    def _write_hb(self, content):
-        if not self.hbfile:
-            return
-        try:
-            _atomic_write(self.hbfile, content)
-        except Exception as e:
-            # The heartbeat drives CARP gating, so a write failure is worth surfacing.
-            LOG.warning("heartbeat write failed (%s): %s", self.hbfile, e)
-
-    def _hb(self):
-        t1, t2, src = self._dhcp.timing()
-        # Publish nudge state so the status page can show it: nudge=<epoch of the
-        # last sent nudge, 0 = never>, arpok=<epoch of the gateway's last ARP reply,
-        # 0 = none seen> and the current target gateway (if known). status.py's
-        # _HB_TOKENS table is the reader -- keep the tokens in lockstep.
-        extra = ""
-        if self._nudge.interval:
-            extra = f" nudge={int(self._nudge.last_nudge)} arpok={int(self._nudge.last_reply)}"
-            gw = self._nudge_target()
-            if gw:
-                extra += f" gw={gw}"
-        self._write_hb(f"{int(time.time())} bound={self._dhcp.binding.yiaddr or '-'} "
-                       f"lease={self._dhcp.binding.lease_secs} t1={t1} t2={t2} src={src}{extra}\n")
-
-    def _hb_mismatch(self, got, want):
-        # Write a clear marker into the heartbeat file so a supervisor/human sees the mismatch.
-        self._write_hb(f"{int(time.time())} MISMATCH got={got} want={want}\n")
-
-    def _on_changed_address(self, got, rx, phase, release_on_enforce):
-        """DhcpClient's changed-address hook -> the FollowPolicy decision. A
-        shim because the client is constructed before the policy exists."""
-        return self._follow.on_changed_address(got, rx, phase, release_on_enforce)
-
-    # ---- CARP role (master probe, transitions) ----
-
-    def _ifconfig(self):
-        """Captured `ifconfig <iface>` text, or None if the probe failed. Shared by
-        the CARP-role probe and the carrier check so the ifconfig invocation, decode
-        and error policy live in exactly one place."""
-        try:
-            out = subprocess.check_output(["/sbin/ifconfig", self.iface], errors="replace")
-        except (OSError, subprocess.SubprocessError):
-            return None
-        return out.decode(errors="replace") if isinstance(out, bytes) else out
-
-    def _probe_carp_master(self):
-        """Raw CARP-role probe for our vhid: True/False from ifconfig, None when
-        the probe itself fails; no vhid configured -> True (nothing to gate on).
-        Callers apply their own policy on a None probe -- the ARP nudge fails closed
-        (no nudge unless a confirmed MASTER); the CARP-transition poll just skips."""
-        if not self.vhid:
-            return True
-        out = self._ifconfig()
-        if out is None:
-            return None
-        return f"carp: MASTER vhid {self.vhid} " in out
-
-    def _carp_master_probe(self):
-        """ArpNudge's is_master hook -> the CARP probe (late-bound through the
-        attribute so tests can stub _probe_carp_master)."""
-        return self._probe_carp_master()
-
-    def _iface_link_up(self):
-        """Interface carrier from ifconfig: True on 'status: active', False on a
-        present-but-inactive status (no carrier / no link), None when it cannot be
-        read (probe failed, or the NIC reports no status line). Used by the
-        unbound link-return fast path and the acquire carrier gate; a None
-        result never disturbs the backoff and never holds an acquire."""
-        out = self._ifconfig()
-        if out is None:
-            return None
-        if "status: active" in out:
-            return True
-        if "status: " in out:
-            return False
-        return None
-
-    def _check_link_returned(self):
-        """While UNBOUND, detect a carrier down->up edge so the keeper re-DORAs at
-        once instead of waiting out the backoff (mirrors dhclient link-up ->
-        state_reboot). Returns True only on a *seen-down* -> up transition, debounced
-        against a flapping link. An initial unknown->up is NOT a trigger (we were
-        already up), so this never fires spuriously at startup."""
-        up = self._iface_link_up()
-        if up is None:
-            return False
-        prev = self._link_up
-        self._link_up = up
-        if up and prev is False:
-            now = time.time()
-            if now - self._link_kick_at < LINK_KICK_DEBOUNCE:
-                return False
-            self._link_kick_at = now
-            return True
-        return False
-
-    def _poll_carp_role(self):
-        """Watch for a backup->master transition (called on the heartbeat
-        cadence). Becoming master renews the lease early and, when enabled,
-        nudges immediately: the failover -- or the link flap that re-elected
-        CARP -- may just have disturbed the upstream gateway's ARP entry and the
-        access node's DHCP-snooping binding, so neither should wait out its
-        normal timer. Independent of the ARP nudge setting."""
-        if not self.vhid:
-            return
-        master = self._probe_carp_master()
-        if master is None:
-            return
-        if master and self._was_master is False:
-            LOG.info("became CARP master for vhid %s -- immediate ARP nudge and early lease renew",
-                     self.vhid)
-            self._renew_asap = True
-            self._arp_nudge(force=True)
-        elif not master and self._was_master:
-            # The symmetric event: without it, "why did the nudges stop?" needs
-            # ifconfig instead of the log.
-            LOG.info("lost CARP master for vhid %s -- ARP nudges pause on this node", self.vhid)
-        self._was_master = master
-
-    # ---- ARP nudge ----
-
-    def _nudge_target(self):
-        """The gateway whose ARP cache the nudge maintains: DHCP option 3 from
-        the last ACK, falling back to the leasing server's address."""
-        return self._dhcp.binding.router or self._dhcp.binding.server
-
-    def _arp_nudge(self, force=False):
-        """Hand the current lease binding (leased address, nudge target) to the
-        ArpNudge component; it owns the pacing, the master gate and the send."""
-        self._nudge.maybe_nudge(self._dhcp.binding.yiaddr, self._nudge_target(), force=force)
-
-    # ---- operator/signal API (set flags only; the loops act within a second) ----
-
-    def request_stop(self):
-        """Ask the daemon to exit at the next loop tick (SIGINT/SIGTERM)."""
-        self._stop = True
-
-    def trigger_nudge(self):
-        """Request an immediate ARP nudge (SIGUSR1 / configd action). Flag
-        only: no network I/O happens in signal context."""
-        self._nudge_now = True
-
-    def recheck_carp_role(self):
-        """Re-check the CARP role within a second (SIGUSR2 from the CARP
-        syshook) instead of waiting for the next ~30s poll."""
-        self._poll_role_now = True
-
-    # ---- main loop / sleeps ----
-
-    def _sleep(self, secs):
-        slept = 0
-        while slept < secs and not self._stop:
-            time.sleep(1)
-            slept += 1
-        return slept
-
-    def _sleep_interruptible(self, secs):
-        """Sleep up to secs (1s steps). Return False early on stop. Also services an
-        operator-requested immediate nudge (SIGUSR1) and a CARP-transition re-check
-        (SIGUSR2) so both act within a second instead of at the next heartbeat tick."""
-        slept = 0
-        while slept < secs and not self._stop:
-            if self._poll_role_now:
-                self._poll_role_now = False
-                # A CARP transition just fired (kernel -> devd -> rc.syshook.d/carp
-                # -> SIGUSR2); re-check the role now so a backup->master keeper
-                # nudges + renews immediately instead of at the next ~30s poll.
-                self._poll_carp_role()
-
-            if self._nudge_now:
-                self._nudge_now = False
-                # Operator actions are rare and intentional -- always log them,
-                # unlike the periodic nudges (whose freshness the status page
-                # already shows without flooding the log every interval).
-                LOG.info("manual ARP nudge requested (SIGUSR1)")
-                self._arp_nudge(force=True)
-                self._hb()   # publish the new nudge age right away for the status page
-
-            # A pending peer-ACK observation follows now (rationale + the
-            # dual-master window it closes: FollowPolicy.check_observed).
-            self._follow.check_observed()
-
-            # Link-return fast path: only while UNBOUND, poll carrier every few
-            # seconds; a down->up edge means the WAN just came back, so stop waiting
-            # and let _maintain_step re-DORA immediately (the bound path skips this).
-            if self._dhcp.binding.yiaddr is None and slept % LINK_POLL_STEP == 0 and self._check_link_returned():
-                self._link_returned = True
-                return not self._stop
-
-            # Event-driven sleep: return at once when the sniffer signals a fresh
-            # observation, otherwise time out after ~1s to run the periodic checks.
-            if self._wake.wait(1.0):
-                self._wake.clear()
-            slept += 1
-        return not self._stop
-
-    def _hold_lease(self, secs):
-        """Sleep up to secs while holding a lease, rewriting the heartbeat every
-        HB_REFRESH so a healthy keeper never looks stale (leases can be hours and
-        the CARP demotion hook only sees heartbeat freshness). Returns False early
-        on stop."""
-        remaining = secs
-        while remaining > 0 and not self._stop:
-            if self._renew_asap:
-                # Return as if T1 elapsed: the caller renews right away, which
-                # re-teaches upstream DHCP-snooping state after a master change.
-                self._renew_asap = False
-                return not self._stop
-
-            chunk = min(HB_REFRESH, remaining)
-            if not self._sleep_interruptible(chunk):
-                return False
-            remaining -= chunk
-
-            self._hb()
-            self._follow.watchdog()   # re-drive a follow whose apply stalled
-            self._poll_carp_role()    # backup->master? renew early + nudge now
-            self._arp_nudge()
-        return not self._stop
-
-    def run(self):
-        """The daemon main loop: capture up, then maintain the lease until
-        stopped. Never raises; returns the process exit code."""
-        # Start the packet capture, retrying forever: a keeper must self-heal, not
-        # die, if the interface is briefly unavailable at startup.
-        while not self._stop and not self._capture.start():
-            LOG.warning("capture start failed -- retrying in %ds", SNIFFER_RETRY)
-            self._sleep(SNIFFER_RETRY)
-
-        # eth-src matters for L2 debugging but only when it differs from the chaddr.
-        ethsrc = f" eth-src={self.eth_src}" if self.eth_src != self.chaddr else ""
-        LOG.info("lease-keeper active on %s: chaddr=%s%s request=%s",
-                 self.iface, self.chaddr, ethsrc, self._follow.target or "any")
-        time.sleep(SNIFFER_WARMUP)
-
-        while not self._stop:
-            # The keeper must NEVER die on a bad DHCP state: catch any unexpected
-            # error, keep the heartbeat fresh so CARP does not falsely demote us,
-            # and retry after a short backoff.
-            try:
-                self._maintain_step()
-            except Exception:
-                LOG.exception("unexpected error in main loop -- recovering")
-                try:
-                    self._hb()
-                except Exception:
-                    pass
-                self._sleep(LOOP_ERROR_BACKOFF)
-
-        # shutdown
-        if self.release_on_exit:
-            self._dhcp.release()
-        self._capture.stop()
-        LOG.info("stopped")
-        return 0
-
-    def _maintain_step(self):
-        """One iteration of the maintain loop. Returns to run() (which loops again)
-        on every state transition; any exception it raises is caught by run() and
-        retried, so a transient fault can never terminate the keeper."""
-        b = self._dhcp.binding
-        if not b.yiaddr:
-            self._acquire_step()
-            return
-
-        # Maintain: wait until T1, then RENEW; bail early on stop.
-        t1, t2, src = self._dhcp.timing()
-        # The renew/rebind plan is verbose and identical every cycle for a stable
-        # lease, so log it at DEBUG -- the default (INFO) log stays clean and
-        # "RENEW ok" already carries the lease + expiry. Raise the keeper's log
-        # level to see it.
-        LOG.debug("DHCP lease %ds; renew at T1=%ds (~%s), rebind by T2=%ds (~%s) (timing source: %s)",
-                  b.lease_secs, t1, _clock_at(t1), t2, _clock_at(t2), src)
-        if not self._hold_lease(t1):
-            return
-        if self._dhcp.renew():
-            LOG.info("DHCP RENEW ok %s (lease=%ss, expires ~%s)",
-                     b.yiaddr, b.lease_secs, _clock_at(b.lease_secs))
-            self._hb()
-            self._arp_nudge(force=True)
-            return
-        LOG.warning("DHCP RENEW failed at T1 -- trying REBIND until T2")
-
-        elapsed = t1
-        ok = False
-        while elapsed < t2 and not self._stop:
-            # Jitter the REBIND retransmit cadence: both nodes hit T2 together
-            # (identical lease timers), so an un-jittered step would broadcast
-            # REBIND in lockstep. Overshooting t2 slightly is harmless (the guard
-            # re-checks). Account the actual jittered wait so elapsed tracks it.
-            step = _jittered(min(REBIND_POLL_STEP, t2 - elapsed))
-            if not self._sleep_interruptible(step):
-                break
-            elapsed += step
-            self._hb()   # still holding the lease until T2 -- keep the heartbeat fresh
-            if self._dhcp.renew(rebind=True):
-                ok = True
-                break
-        if ok:
-            LOG.info("DHCP REBIND ok %s (lease=%ss, expires ~%s)",
-                     b.yiaddr, b.lease_secs, _clock_at(b.lease_secs))
-            self._hb()
-            self._arp_nudge(force=True)
-            return
-
-        LOG.error("DHCP lease expired -- re-acquiring (back to DISCOVER)")
-        self._dhcp.expire()
-
-    def _acquire_step(self):
-        """The unbound arm of the maintain loop: hold while there is no
-        carrier, otherwise try one acquire with the jittered re-DORA backoff;
-        the link-return fast path cuts either wait short."""
-        self._ensure_sniffer()  # a dead sniffer would silently fail every DORA
-        self._hb()  # active but not holding yet -> publish bound=-
-
-        # No point broadcasting DISCOVERs on a dead link (native dhclient
-        # does not either): wait for the carrier instead of burning DORA
-        # attempts. The unbound sleep polls the link and the link-return
-        # fast path resumes within seconds of it coming back. Fail open:
-        # only a confirmed "no carrier" (False) holds the acquire; an
-        # unreadable probe (None) never blocks it.
-        if self._iface_link_up() is False:
-            if self._link_up is not False:   # log once per down-episode
-                LOG.info("no carrier on %s -- waiting for the link before the DHCP acquire",
-                         self.iface)
-            self._link_up = False
-            self._link_returned = False
-            self._sleep_interruptible(REDORA_MIN)
-            if self._link_returned:
-                LOG.info("WAN link returned while unbound -- re-acquiring now")
-                self.redora_wait = REDORA_MIN
-            return
-
-        b = self._dhcp.binding
-        if self._dhcp.acquire(self._follow.target):
-            self._link_up = True   # a completed DORA proves carrier
-            LOG.info("DHCP BOUND %s (lease=%ss, expires ~%s, server=%s, gw=%s, mask=%s)",
-                     b.yiaddr, b.lease_secs, _clock_at(b.lease_secs),
-                     b.server, b.router or "?",
-                     f"/{b.mask_bits}" if b.mask_bits else "none")
-            self._hb()
-            self._arp_nudge(force=True)
-            self.redora_wait = REDORA_MIN
-        else:
-            LOG.warning("DHCP acquire (DISCOVER/REQUEST) failed -- retrying in %ds", self.redora_wait)
-            self._link_returned = False
-            self._sleep_interruptible(_jittered(self.redora_wait))
-            if self._link_returned:
-                # WAN carrier returned mid-backoff: re-acquire now instead of
-                # waiting out the next backoff (matches native dhclient).
-                LOG.info("WAN link returned while unbound -- re-acquiring now")
-                self.redora_wait = REDORA_MIN
-            else:
-                self.redora_wait = min(self.redora_wait * 2, REDORA_MAX)
-
-
-def _identity_options(vendor_class, client_id, hostname):
-    """Optional DHCP identity options (empty -> not sent), added to every
-    DISCOVER/REQUEST/RENEW so the server sees a consistent client identity.
-    ISP interplay: satisfies servers that only lease to a known vendor-class
-    (opt 60), client-id (61) or hostname (12) -- the "client identity checks"
-    row of the README's ISP-security section."""
-    id_opts = []
-    if vendor_class:
-        id_opts.append(("vendor_class_id", vendor_class))
-    if client_id:
-        id_opts.append(("client_id", client_id.encode()))
-    if hostname:
-        id_opts.append(("hostname", hostname))
-    return id_opts
-
+from leasekeeper.keeper import Keeper
+# Transitional re-exports (removed when the tests import from leasekeeper.*):
+from leasekeeper.dhcpclient import DhcpClient, Lease  # noqa: F401  pylint: disable=unused-import
+from leasekeeper.policy import ArpNudge, FollowPolicy  # noqa: F401  pylint: disable=unused-import
 
 def acquire_pidfile(path):
     """Single-instance guard: atomically claim the pidfile, replacing a stale
